q1/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.10.1
torch==2.0.1
transformers==4.31.0
sentence-transformers==2.2.2
gensim==4.3.1
streamlit==1.24.1
plotly==5.15.0
openai==0.27.8
python-dotenv==1.0.0
nltk==3.8.1
umap-learn==0.5.3
huggingface-hub==0.16.4 